apps/ai-service/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pytesseract
pdf2image
Pillow
transformers
torch
python-dotenv
google-generativeai
soxr
